+++ deltachat-ffi/deltachat.h
@@ -407,9 +407,8 @@ char*           dc_get_blobdir               (const dc_context_t* context);
  *                    to not mess up with non-delivery-reports or read-receipts.
  *                    0=no limit (default).
  *                    Changes affect future messages only.
- * - `debug_logging` = 1=Let the core save all events to the database. You should expose this as an advanced
- *                    setting to the user.
- *                    0=disable debug logging (default)
+ * - `debug_logging` = 1: Let the core save all events to the database.
+ *                     0: disable debug logging and delete logging xdc (default)
  * - `ui.*`         = All keys prefixed by `ui.` can be used by the user-interfaces for system-specific purposes.
  *                    The prefix should be followed by the system and maybe subsystem,
  *                    e.g. `ui.desktop.foo`, `ui.desktop.linux.bar`, `ui.android.foo`, `ui.dc40.bar`, `ui.bot.simplebot.baz`.