+++ CHANGELOG.md
@@ -58,7 +58,7 @@
 - Remove `pytest-async` plugin #3846
 - Only send the message about ephemeral timer change if the chat is promoted #3847
 - Use relative paths in `accounts.toml` #3838
-- add `debug_logging` setting #3296
+- add debug logging support for webxdcs #3296
 
 ### Fixes
 - Set read/write timeouts for IMAP over SOCKS5 #3833
@@ -90,7 +90,6 @@
 - Add Python API to send reactions #3762
 - jsonrpc: add message errors to MessageObject #3788
 - jsonrpc: Add async Python client #3734
-- add `debug_logging` setting #3296
 
 ### Fixes
 - Make sure malformed messsages will never block receiving further messages anymore #3771